refactor(voc): fix slice structure — constants, useVocPermissions, review/index
- Move list/constants.ts → voc/constants.ts (used across list + review)
- Move voc/model/useVocPermissions.ts → review/model/useVocPermissions.ts
  (review-only hook belongs in review slice)
- Add review/index.ts exporting VocReviewDrawer (parity with other slices)
- Update all import paths accordingly
