+++ frontend/src/features/voc/constants.ts
