+++ frontend/src/features/voc/list/index.ts
@@ -4,5 +4,5 @@ export { useVocFilter } from './model/useVocFilter';
 export { useVocFilters } from './model/useVocFilters';
 export type { VocPageFilters } from './model/useVocFilters';
 export { VocListHeader } from './ui/VocListHeader';
-export { VOC_FIELD_LABELS } from './constants';
+export { VOC_FIELD_LABELS } from '../constants';
 export { VOC_GRID_COLUMNS, VOC_GRID_PADDING_X } from './ui/vocGridLayout';
