+++ frontend/src/features/voc/review/index.ts
@@ -1,0 +1,1 @@
+export { VocReviewDrawer } from './ui/VocReviewDrawer';
